requirements.txt
python-binance
python-dotenv
ta
scikit-learn
pandas
